Update README.md
In this project,  will make use of Python to explore data related to bike share systems for three major cities in the United States—Chicago, New York City, and Washington. Will write code to import the data and answer interesting questions about it by computing descriptive statistics. Will also write a script that takes in raw input to create an interactive experience in the terminal to present these statistics.